ADD - ASStructure.java
