+++ src/interpreteur/as/lang/datatype/ASStructure.java
@@ -1,0 +1,22 @@
+package interpreteur.as.lang.datatype;
+
+public class ASStructure implements ASObjet<Object> {
+
+    public ASStructure() {
+    }
+
+    @Override
+    public Object getValue() {
+        return null;
+    }
+
+    @Override
+    public boolean boolValue() {
+        return false;
+    }
+
+    @Override
+    public String obtenirNomType() {
+        return null;
+    }
+}
